+++ toduck/Source/Presentation/Home/Toduck/ToduckViewController.swift
@@ -10,27 +10,8 @@ import SnapKit
 import Then
 
 final class ToduckViewController: UIViewController {
-    
-    let calendarView = UICalendarView().then {
-        $0.backgroundColor = .systemBackground
-        $0.tintColor = TDColor.Diary.angry
-    }
-    
     override func viewDidLoad() {
         super.viewDidLoad()
-        
-        setCalendarView()
-    }
-}
-
-extension ToduckViewController {
-    private func setCalendarView() {
-        view.addSubview(calendarView)
-        
-        calendarView.translatesAutoresizingMaskIntoConstraints = false
-        calendarView.snp.makeConstraints {
-            $0.centerX.equalTo(view)
-            $0.width.equalTo(view.snp.width).multipliedBy(0.9)
-        }
+        view.backgroundColor = .blue
     }
 }
